Implement q_size function
Implement the q_size function to return the number of elements in a queue
represented by a linked list. The function iterates through the queue using
a `list_for_each` loop and increments a counter to determine the size of
the queue. If the input head is NULL, the function returns 0 to indicate
an empty queue.